Fix initial migration, plus some simplification. (#23)
The migrate command was failing for new installs, mainly
because the code for creating the initial game was borked
due to a rename. Add a migrate command in the GitHub action
to try to ensure it stays working.

Additionally, the multi-site setup was cumbersome, so it's
been simplified for now. Site autodetection isn't used, instead
whatever game is indicated by GAME_ID is the game. If multiple
game sites are needed, run a binary for each with a different
environment variable set.

The game hub is also removed for now to keep it simple.