Add Artisan (#101)
* Add Artisan base definition, archetype, and innate powers.

* Artisan Skills

* Add all Artian powers.

* Update CP cost rules.

* Validate and display feature tags.

Additionally, minor formatting fixes for existing descriptions.
I had been using `>` for multiline string descriptions, but that
replaces all newlines with spaces. Replacing with `|` causes newlines
to be preserved, also removing the need for the hacky `<br/>` tags
I had been using between paragraphs.

* Support buying martial (tiered) powers from your class.

This includes the ability to purchase low-tier powers with
a high-tier slot if available.

* Add support for utility powers.

* Implement artisan specializations, part 1.

This covers calculation of the current specialization,
including a tiebreaker choice if the top tags are
equally matched.

It does not yet expose this information for use by
features that might rely on it.

* Implement Opus.

This allows `artisan.specialization+crafter` etc to be used as
a requirement. The value of these expressions will be the number
of powers held with that tag, but only if that tag is the current
specialization.

Also adds a `grant_if` field for Tempest feature definitions
of the form:

```yaml
grant_if:
    feature-a: requirements-for-a
    feature-b: requirements-for-b
```