Add breed support, Human, and Thrylossi. (#119)
* Add breed schema and Human files.

* Add breed, subbreed, and breed challenge controllers.

Fix some bugs related to attribute fetching.

* whoops missed a file...

* Remove the obligatory definition type check from each feature controller.

This is now performed in the base constructor, but only in __debug__.

* First draft of Advantages support.

Note to self, this messes with subfeature purchase rendering.
Go back and look at Right Hand to see if it looks weird.

* Add Thrylossi.

Revisit subfeature purchase displays.

Return-to-parent behavior updated. This will
only happen for 'internal' types. Not that if a feature
has purchases available, it will default to sticking to
itself, just as it would for Choices available.

Fixes #86