Display cost (or award) values on overview. (#127)
These are displayed via a badge. For items that normally would
display a cost but are free, no badge will be displayed. For items
that are normally "free" like powers and spells, no badge will
be displayed unless the feature is granted, in which case a "Granted"
badge will be shown.

This should probably largely address #62 and #65.