Use a blank choice definition by default.
Some choice contollers aren't the result of a choice
definition and are instead the result of some other
mechanic, but the base class refers to the definition
in several places. Consider splitting this into an
interface in the future, but for now just provide
a blank definition if needed.

Fixes #146