Allow profile data to be created/updated on event registration.
This adds a collapse to the event registration with a sub-form.
Depending on whether the profile data was already filled in or
not, the collapse will be expanded or collapsed by default.

Normally if you try to submit a form and there's a client-side validation issue
like "I didn't fill in my legal name", the browser will scroll there
and focus it in a passive aggressive manner. But if the fields are inside of
a collapsed collapse, by default the browser doesn't know what to do, so it just
fails and shrugs. Wiring up something to those elements' `invalid` event to pop
the collapse open seems to do the trick.

Fixes #176.