Refinements for event views.
Events can be marked complete (if done in the right order) and
completed (or canceled) events in the past will not display
by default in the events list.

When an event is marked complete, it is integrated into the
campaign model (that is, the Max XP and Max CP go up if appropriate).

The registration view is cleaned up in prep for award assignment work.