Add a management command for recalculating campaign data.
This includes the ability to force a particular event to be marked completed,
even if it would otherwise be prevented in the UI due to sequence breaking.

This also updates the CampaignRecord to record all event data in case this
is wanted later. As a result, the recent_events list is no longer needed,
and ahistorical events *could* be added - but we'll continue preventing
this in the UI for now.

Also, noticed a bug related to adding a Date object into the character
metadata - the dict_dump command used for character serialization had
not been updated to use the mode='json' flag, so this would fail.

Added an integration test that would have caught this.